Fixed z-fighting and changed angle math a bit to look better
Also merge into master since this is the last commit for a while. As
such, I've removed all but two debug messages about loading which are
useful to the client player. Also added the glow images in the event
that when I am long gone, imgur is down (or is just laggy for certain
players) so that you may rehost. Images were created by me and they are
free to use, distribute, sell, whatever you want to do with them
(inspired by George Lucas' creations)