Fixed rejoining bug
If a player had changed their lightsaber, upon rejoining the module would fail to load any lightsaber at all. Causes are to be speculated on but it's probably something to do with the Player class being a bit weird when the script is run.

Reminder: this bug and many more were fixed on the Streamed branch.